+++ app/src/main/java/ru/netology/nmedia/MainActivity.kt
@@ -26,23 +26,9 @@ class MainActivity : ComponentActivity() {
         binding.root.adapter = adapter
         viewModel.data.observe(this) {posts ->
             adapter.submitList(posts)
-
         }
-
-    }
-
-}
-
-
-fun countViews(count:Long): String{
-    val array = arrayOf(' ', 'k', 'M', 'B', 'T', 'P', 'E')
-    val value = floor(log10(count.toDouble())).toInt()
-    val  base = value / 3
-    return if (value >= 3 && base < array.size) {
-        DecimalFormat("#0.0").format(count/ 10.0.pow((base * 3).toDouble())) + array[base]
-    } else {
-        DecimalFormat("#,##0").format(count)
     }
 }
+
