+++ app/src/main/res/layout/post.xml
@@ -33,6 +33,7 @@
         />
 
     <TextView
+        android:id="@+id/published"
         android:layout_width="0dp"
         android:layout_height="wrap_content"
         app:layout_constraintStart_toEndOf="@id/avatar"
@@ -40,7 +41,6 @@
         app:layout_constraintBottom_toBottomOf="@id/avatar"
         app:layout_constraintEnd_toStartOf="@id/more_button"
         android:layout_marginStart="@dimen/common_margin_spacing"
-        android:id="@+id/published"
         android:ellipsize="end"
         android:singleLine="true"
         android:text="@string/published"
@@ -48,8 +48,8 @@
         />
 
     <TextView
-        android:layout_width="match_parent"
         android:id="@+id/post_text"
+        android:layout_width="match_parent"
         android:layout_height="wrap_content"
         app:layout_constraintTop_toBottomOf="@id/header_barrier"
         android:layout_marginTop="@dimen/common_margin_spacing"
@@ -58,6 +58,7 @@
         />
 
     <ImageButton
+        android:id="@+id/more_button"
         android:layout_width="24dp"
         android:layout_height="24dp"
         app:layout_constraintEnd_toEndOf="parent"
@@ -66,21 +67,20 @@
         android:src="@drawable/ic_baseline_more_vert_24"
         app:srcCompat="@drawable/ic_baseline_more_vert_24"
         android:visibility="visible"
-        android:id="@+id/more_button"
         android:background="@android:color/transparent"
         android:contentDescription="@string/show_more_button" />
 
     <androidx.constraintlayout.widget.Barrier
+        android:id="@+id/header_barrier"
         android:layout_width="0dp"
         android:layout_height="0dp"
-        android:id="@+id/header_barrier"
         app:barrierDirection="bottom"
         app:constraint_referenced_ids="avatar,published,more_button"
         />
     <androidx.constraintlayout.widget.Barrier
+        android:id="@+id/post_text_barrier"
         android:layout_width="0dp"
         android:layout_height="0dp"
-        android:id="@+id/post_text_barrier"
         app:barrierDirection="bottom"
         app:constraint_referenced_ids="post_text"
         />
@@ -92,6 +92,7 @@
         android:visibility="visible"
         android:background="@android:color/transparent"
         tools:src="@drawable/ic_like_border_24"
+        app:srcCompat="@drawable/ic_like_border_24"
         app:layout_constraintTop_toTopOf="@id/post_text_barrier"
         android:layout_marginTop="@dimen/common_margin_spacing"
         android:contentDescription="@string/like_icon"
